TEMP: move the GPU diagnostic into the failing test binary
cargo test stops at the first failing binary, so the diagnostic in
brush-cube never ran. Tests within one binary all run, so put it here.
